NoteSequence | cleanup + refactor
