+++ include/config/notesequence.hpp
@@ -15,8 +15,8 @@
 #include "imgui_internal.h"
 #include "ImSequencer.h"
 
+#include "config/constants.hpp"
 #include "config/note.hpp"
-#include "config/notemaps.hpp"
 #include "config/skip.hpp"
 #include "config/stop.hpp"
 #include "config/timeinfo.hpp"
@@ -26,28 +26,9 @@
 
 #include "systems/audiosystem.hpp"
 
-static const char* SequencerItemTypeNames[] = { "Lane 1 [Top]", "Lane 2 [Middle]", "Lane 3 [Bottom]", "Stops", "Skips" };
-
-const std::unordered_map<int, NoteSplit> DENOMINATOR_TO_NOTESPLIT = {
-    { 1, NoteSplit::WHOLE },
-    { 2, NoteSplit::HALF },
-    { 4, NoteSplit::QUARTER },
-    { 8, NoteSplit::EIGHTH },
-    { 16, NoteSplit::SIXTEENTH },
-    { 32, NoteSplit::THIRTYSECOND },
-    { 64, NoteSplit::SIXTYFOURTH },
-    { 128, NoteSplit::ONETWENTYEIGHTH },
-    { 3, NoteSplit::TWELFTH },
-    { 6, NoteSplit::TWELFTH },
-    { 12, NoteSplit::TWELFTH },
-    { 24, NoteSplit::TWENTYFOURTH },
-    { 48, NoteSplit::FORTYEIGHTH },
-    { 96, NoteSplit::NINETYSIXTH }
-};
-
 struct NoteSequence : public ImSequencer::SequenceInterface {
     NoteSequence() = default;
-    ~NoteSequence() = default;
+    virtual ~NoteSequence() = default;
 
     float mFrameMin { 0.f };
     float mFrameMax { 1000.f };
@@ -59,501 +40,51 @@ struct NoteSequence : public ImSequencer::SequenceInterface {
     std::map<std::string, int> keyFrequencies;
     std::vector<std::pair<std::string, int>> keyFreqsSorted;
 
-    NoteSplit computeNoteSplit(int noteSplitNumerator, int noteSplitDenominator) {
-        auto gcd = std::gcd(noteSplitNumerator, noteSplitDenominator);
-        noteSplitNumerator /= gcd;
-        noteSplitDenominator /= gcd;
+    void update(double songBeat, AudioSystem * audioSystem, bool notesoundEnabled);
+    void resetPassed(double songBeat);
 
-        if(DENOMINATOR_TO_NOTESPLIT.find(noteSplitDenominator) != DENOMINATOR_TO_NOTESPLIT.end()) {
-            return DENOMINATOR_TO_NOTESPLIT.at(noteSplitDenominator);
-        } else {
-            return NoteSplit::WHOLE;
-        }    
-    }
+    void addNote(double absBeat, double songBeat, double beatDuration, BeatPos beatpos, BeatPos endBeatpos,
+        NoteSequenceItem::SequencerItemType itemType, std::string_view displayText);
+    void editNote(double absBeat, NoteSequenceItem::SequencerItemType itemType, std::string_view displayText);
 
-    void resetPassed(double songBeat) {
-        for(auto & item : myItems) {
-            switch(item->getItemType()) {
-                case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                    item->passed = item->absBeat < songBeat;
-                    break;
-                case NoteSequenceItem::SequencerItemType::SKIP:
-                    break;
-                case NoteSequenceItem::SequencerItemType::STOP:
-                    break;
-            }
-        }
-    }
+    void addStop(double absBeat, double songBeat, double beatDuration, BeatPos beatpos, BeatPos endBeatpos);
 
-    void update(double songBeat, AudioSystem * audioSystem, bool notesoundEnabled) {
-        for(auto & item : myItems) {
-            switch(item->getItemType()) {
-                case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                    if(item->passed) {
-                        break;
-                    } else if(item->absBeat < songBeat) {
-                        item->passed = true;
+    std::shared_ptr<Skip> addSkip(double absBeat, double songBeat, double skipTime, double beatDuration, BeatPos beatpos, BeatPos endBeatpos);
+    void editSkip(double absBeat, double skipTime);
 
-                        if(notesoundEnabled) {
-                            audioSystem->playSound("keypress");
-                        }
-                    }
-                    break;
-                case NoteSequenceItem::SequencerItemType::SKIP:
-                    break;
-                case NoteSequenceItem::SequencerItemType::STOP:
-                    break;
-            }
-        }
-    }
-
-    static bool cmpSecond(const std::pair<std::string, int> & l, const std::pair<std::string, int> & r) {
-        return l.second > r.second;
-    }
-
-    void addNote(double absBeat, double songBeat, double beatDuration, BeatPos beatpos, BeatPos endBeatpos, NoteSequenceItem::SequencerItemType itemType, std::string displayText) {
-        NoteType noteType = NoteType::KEYPRESS;
-        if(beatDuration > FLT_EPSILON) {
-            noteType = NoteType::KEYHOLDSTART;
-        }
-
-        bool passed = absBeat < songBeat;
-
-        std::shared_ptr<NoteSequenceItem> newNote = std::make_shared<Note>(absBeat, absBeat + beatDuration, passed, beatpos, endBeatpos, 
-            noteType, NoteSplit::EIGHTH, itemType, displayText);
-        myItems.push_back(newNote);
-
-        std::sort(myItems.begin(), myItems.end());
-
-        switch(itemType) {
-            case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                numTopNotes++;
-                break;
-            case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                numMidNotes++;
-                break;
-            case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                numBotNotes++;
-                break;
-            default:
-                break;
-        }
-
-        keyFrequencies[displayText] += 1;
-
-        updateKeyFrequencies();
-    }
-
-    void flipNotes(std::string_view keyboardLayout, double startBeat, double endBeat, int minItemType, int maxItemType) {
-        if(notemaps::KEYBOARD_FLIP_MAPS.find(keyboardLayout) != notemaps::KEYBOARD_FLIP_MAPS.end()) {
-            auto & flipMap = notemaps::KEYBOARD_FLIP_MAPS.at(keyboardLayout);
-
-            auto items = getItems(startBeat, endBeat, minItemType, maxItemType);
-            for(auto item: items) {
-                switch(item->getItemType()) {
-                    case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                    case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                    case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                        if(flipMap.find(item->displayText) != flipMap.end()) {
-                            keyFrequencies[item->displayText] -= 1;
-                            item->displayText = flipMap.at(item->displayText);
-                            keyFrequencies[item->displayText] += 1;
-                        }
-                        break;
-                    default:
-                        break;
-                }
-            }
-
-            updateKeyFrequencies();
-        }
-    }
+    void flipNotes(std::string_view keyboardLayout, double startBeat, double endBeat, int minItemType, int maxItemType);
 
     std::list<std::shared_ptr<NoteSequenceItem>> shiftNotes(std::string_view keyboardLayout, double startBeat, double endBeat,
-        int minItemType, int maxItemType, ShiftNoteAction::ShiftDirection shiftDirection)
-    {
-        auto items = getItems(startBeat, endBeat, minItemType, maxItemType);
-        return shiftItems(keyboardLayout, startBeat, endBeat, items, shiftDirection);
-    }
-
+        int minItemType, int maxItemType, ShiftNoteAction::ShiftDirection shiftDirection);
     std::list<std::shared_ptr<NoteSequenceItem>> shiftItems(std::string_view keyboardLayout, double startBeat, double endBeat,
-        const std::list<std::shared_ptr<NoteSequenceItem>> & items, ShiftNoteAction::ShiftDirection shiftDirection)
-    {
-        std::list<std::shared_ptr<NoteSequenceItem>> shiftedItems;
+        const std::list<std::shared_ptr<NoteSequenceItem>> & items, ShiftNoteAction::ShiftDirection shiftDirection);
+    bool shiftNoteSequenceItem(ShiftNoteAction::ShiftDirection shiftDirection, std::shared_ptr<NoteSequenceItem> item, std::string_view keyboardLayout);
 
-        if (notemaps::KEYBOARD_LAYOUTS.find(keyboardLayout) != notemaps::KEYBOARD_LAYOUTS.end() &&
-            notemaps::KEYBOARD_POSITION_MAPS.find(keyboardLayout) != notemaps::KEYBOARD_POSITION_MAPS.end())
-        {
-            auto & keyboardLayoutMap = notemaps::KEYBOARD_LAYOUTS.at(keyboardLayout);
-            auto & keyboardPositionMap = notemaps::KEYBOARD_POSITION_MAPS.at(keyboardLayout);
+    std::list<std::shared_ptr<NoteSequenceItem>> getItems(double startBeat, double endBeat, int minItemType, int maxItemType);
+    std::shared_ptr<NoteSequenceItem> containsItemAt(double absBeat, NoteSequenceItem::SequencerItemType itemType);
+    int getLaneItemCount(NoteSequenceItem::SequencerItemType lane) const;
+    void resetItemCounts();
 
-            for(auto item: items) {
-                auto shifted = shiftNoteSequenceItem(shiftDirection, item, keyboardLayoutMap, keyboardPositionMap);
-                if(shifted) {
-                    shiftedItems.push_back(item);
-                }
-            }
+    void insertItems(double insertBeat, double songBeat, int minItemType, int maxItemType, const std::vector<Timeinfo> & timeinfo, std::list<std::shared_ptr<NoteSequenceItem>> items);
+    void deleteItems(double startBeat, double endBeat, int minItemType, int maxItemType);
+    void deleteItem(double absBeat, NoteSequenceItem::SequencerItemType itemType);
 
-            updateKeyFrequencies();
-        }
-
-        return shiftedItems;
-    }
-
-    bool shiftNoteSequenceItem(ShiftNoteAction::ShiftDirection shiftDirection, std::shared_ptr<NoteSequenceItem> item,
-        const std::vector<std::vector<std::string>> & keyboardLayoutMap, const std::unordered_map<std::string, std::pair<int, int>> & keyboardPositionMap)
-    {
-        auto itemKey = item->displayText;
-        auto itemType = item->getItemType();
-
-        switch(itemType) {
-            case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-            case NoteSequenceItem::SequencerItemType::MID_NOTE:
-            case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                if(keyboardPositionMap.find(itemKey) != keyboardPositionMap.end()) {
-                    auto keyPos = keyboardPositionMap.at(itemKey);
-                    int keyRow = keyPos.first;
-                    int keyCol = keyPos.second;
-
-                    int newRow = keyRow;
-                    int newCol = keyCol;
-                    switch(shiftDirection) {
-                        case ShiftNoteAction::ShiftDirection::ShiftUp:
-                            newRow = keyRow - 1;
-                            break;
-                        case ShiftNoteAction::ShiftDirection::ShiftDown:
-                            newRow = keyRow + 1;
-                            break;
-                        case ShiftNoteAction::ShiftDirection::ShiftLeft:
-                            newCol = keyCol - 1;
-                            break;
-                        case ShiftNoteAction::ShiftDirection::ShiftRight:
-                            newCol = keyCol + 1;
-                            break;
-                        default:
-                            break;
-                    }
-
-                    newRow = std::max(0, std::min(newRow, 3));
-                    newCol = std::max(0, std::min(newCol, 9));
-
-                    if(newRow == keyRow && newCol == keyCol) {
-                        return false;
-                    }
-
-                    if(itemType == NoteSequenceItem::SequencerItemType::TOP_NOTE && newRow > 0) {
-                        item->setItemType(NoteSequenceItem::SequencerItemType::MID_NOTE);
-                    }
-
-                    if(itemType == NoteSequenceItem::SequencerItemType::MID_NOTE && newRow < 1) {
-                        item->setItemType(NoteSequenceItem::SequencerItemType::TOP_NOTE);
-                    }
-
-                    auto newKey = keyboardLayoutMap[newRow][newCol];
-                    keyFrequencies[itemKey] -= 1;
-                    item->displayText = newKey;
-                    keyFrequencies[newKey] += 1;
-                }
-                break;
-            default:
-                break;
-        }
-
-        return true;
-    }
-
-    void addStop(double absBeat, double songBeat, double beatDuration, BeatPos beatpos, BeatPos endBeatpos) {
-        bool passed = absBeat < songBeat;
-        auto newStop { std::make_shared<Stop>(absBeat, beatDuration, passed, beatpos, endBeatpos) };
-        newStop->displayText = std::to_string(beatDuration);
-        myItems.push_back(newStop);
-
-        std::sort(myItems.begin(), myItems.end());
-    }
-
-    std::shared_ptr<Skip> addSkip(double absBeat, double songBeat, double skipTime, double beatDuration, BeatPos beatpos, BeatPos endBeatpos) {
-        bool passed = absBeat < songBeat;
-        auto newSkip { std::make_shared<Skip>(absBeat, skipTime, passed, beatDuration, beatpos, endBeatpos) };
-        newSkip->displayText = std::to_string(skipTime);
-        myItems.push_back(newSkip);
-
-        std::sort(myItems.begin(), myItems.end());
-
-        return newSkip;
-    }
-
-    void editSkip(double absBeat, double skipTime) {
-        for(auto iter = myItems.begin(); iter != myItems.end(); iter++) {
-            auto & seqItem = *iter;
-            if(seqItem->getItemType() == NoteSequenceItem::SequencerItemType::SKIP && absBeat >= seqItem->absBeat &&
-                (absBeat < seqItem->beatEnd || (seqItem->absBeat == seqItem->beatEnd && absBeat <= seqItem->beatEnd))) {
-
-                auto currSkip = std::dynamic_pointer_cast<Skip>(seqItem);
-                currSkip->displayText = std::to_string(skipTime);
-                currSkip->skipTime = skipTime;
-        
-                break;
-            }
-        }
-    }
-
-    void editNote(double absBeat, int itemType, std::string displayText) {
-        for(auto iter = myItems.begin(); iter != myItems.end(); iter++) {
-            auto & seqItem = *iter;
-            if((int)(seqItem->getItemType()) == itemType && absBeat >= seqItem->absBeat &&
-                (absBeat < seqItem->beatEnd || (seqItem->absBeat == seqItem->beatEnd && absBeat <= seqItem->beatEnd))) {
-                
-                auto oldText = seqItem->displayText;
-                seqItem->displayText = displayText;
-
-                keyFrequencies[oldText] -= 1;
-                keyFrequencies[displayText] += 1;
-
-                updateKeyFrequencies();
-
-                break;
-            }
-        }
-    }
-
-    std::list<std::shared_ptr<NoteSequenceItem>> getItems(double startBeat, double endBeat, int minItemType, int maxItemType) {
-        std::list<std::shared_ptr<NoteSequenceItem>> currItems;
-
-        for(auto iter = myItems.begin(); iter != myItems.end(); iter++) {
-            auto & seqItem = *iter;
-
-            int seqItemType = (int)(seqItem->getItemType());
-
-            if(seqItemType >= minItemType && seqItemType <= maxItemType && startBeat <= seqItem->absBeat && seqItem->absBeat <= endBeat) {
-                currItems.push_back(seqItem);
-            } else if(endBeat <= seqItem->absBeat) {
-                break;
-            }
-        }
-
-        return currItems;
-    }
-
-    void insertItems(double insertBeat, double songBeat, int minItemType, int maxItemType,
-        const std::vector<Timeinfo> & timeinfo, std::list<std::shared_ptr<NoteSequenceItem>> items)
-    {
-        if(!items.empty()) {
-            double firstBeat = items.front()->absBeat;
-            BeatPos firstBeatPos = items.front()->beatpos;
-            BeatPos insertBeatPos = utils::calculateBeatpos(insertBeat, firstBeatPos.measureSplit, timeinfo);
-
-            deleteItems(insertBeat, insertBeat + (items.back()->beatEnd - firstBeat), minItemType, maxItemType);
-            
-            for(auto item : items) {
-                double currBeat = insertBeat + (item->absBeat - firstBeat);
-
-                BeatPos currBeatPos = insertBeatPos + (item->beatpos - firstBeatPos);
-                BeatPos currEndBeatPos = insertBeatPos + (item->endBeatpos - firstBeatPos);
-
-                switch(item->getItemType()) {
-                    case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                    case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                    case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                        addNote(currBeat, songBeat, item->beatEnd - item->absBeat, currBeatPos, currEndBeatPos, item->getItemType(), item->displayText);
-                        break;
-                    case NoteSequenceItem::SequencerItemType::STOP:
-                        addStop(currBeat, songBeat, item->beatEnd - item->absBeat, currBeatPos, currEndBeatPos);
-                        break;
-                    case NoteSequenceItem::SequencerItemType::SKIP:
-                        auto currSkip = std::dynamic_pointer_cast<Skip>(item);
-                        addSkip(currBeat, songBeat, currSkip->skipTime, item->beatEnd - item->absBeat, currBeatPos, currEndBeatPos);
-                        break;
-                }
-            }
-        }
-    }
-
-    void deleteItems(double startBeat, double endBeat, int minItemType, int maxItemType) {
-        for(auto iter = myItems.begin(); iter != myItems.end();) {
-            bool removeNote = false;
-            const auto & seqItem = *iter;
-
-            auto seqItemType = static_cast<int>(seqItem->getItemType());
-
-            if(seqItemType >= minItemType && seqItemType <= maxItemType && startBeat <= seqItem->absBeat && seqItem->absBeat <= endBeat) {
-                switch(seqItem->getItemType()) {
-                    case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                        numTopNotes--;
-                        removeNote = true;
-                        break;
-                    case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                        numMidNotes--;
-                        removeNote = true;
-                        break;
-                    case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                        numBotNotes--;
-                        removeNote = true;
-                        break;
-                    default:
-                        break;
-                }
-
-                if(removeNote) {
-                    keyFrequencies[seqItem->displayText] -= 1;
-                    if(keyFrequencies[seqItem->displayText] == 0) {
-                        keyFrequencies.erase(seqItem->displayText);
-                    }
-                }
-
-                // in case dangling pointers in undo/redo stack refer to this item
-                seqItem->deleted = true;
-
-                iter = myItems.erase(iter);
-            } else {
-                iter++;
-            }
-        }
-
-        updateKeyFrequencies();
-    }
-
-    void deleteItem(double absBeat, NoteSequenceItem::SequencerItemType itemType) {
-        bool removeNote = false;
-        for(auto iter = myItems.begin(); iter != myItems.end(); iter++) {
-            const auto & seqItem = *iter;
-            if(seqItem->getItemType() == itemType && absBeat >= seqItem->absBeat &&
-                (absBeat < seqItem->beatEnd || (seqItem->absBeat == seqItem->beatEnd && absBeat <= seqItem->beatEnd))) {
-                switch(seqItem->getItemType()) {
-                    case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                        numTopNotes--;
-                        removeNote = true;
-                        break;
-                    case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                        numMidNotes--;
-                        removeNote = true;
-                        break;
-                    case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                        numBotNotes--;
-                        removeNote = true;
-                        break;
-                    default:
-                        break;
-                }
-
-                if(removeNote) {
-                    keyFrequencies[seqItem->displayText] -= 1;
-
-                    if(keyFrequencies[seqItem->displayText] == 0) {
-                        keyFrequencies.erase(seqItem->displayText);
-                    }
-
-                    updateKeyFrequencies();
-                }
-
-                // in case dangling pointers in undo/redo stack refer to this item
-                seqItem->deleted = true;
-
-                myItems.erase(iter);
-                break;
-            }
-        }
-    }
-
-    std::shared_ptr<NoteSequenceItem> containsItemAt(double absBeat, int itemType) {
-        for(const auto & seqItem : myItems) {
-            if((int)(seqItem->getItemType()) == itemType && absBeat >= seqItem->absBeat &&
-                (absBeat < seqItem->beatEnd || (seqItem->absBeat == seqItem->beatEnd && absBeat <= seqItem->beatEnd))) {
-                return seqItem;
-            }
-        }
-
-        return nullptr;
-    }
-
-    int getLaneItemCount(NoteSequenceItem::SequencerItemType lane) const {
-        switch(lane) {
-            case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                return numTopNotes;
-            case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                return numMidNotes;
-            case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                return numBotNotes;
-            default:
-                return 0;
-        }
-    }
-
-    std::pair<std::string, int> & getKeyItemData(int frequencyRank) {
-        return keyFreqsSorted.at(frequencyRank);
-    }
-
-    void resetItemCounts() {
-        numTopNotes = 0;
-        numMidNotes = 0;
-        numBotNotes = 0;
-
-        for(auto & item : myItems) {
-            bool insertedKey = false;
-
-            switch(item->getItemType()) {
-                case NoteSequenceItem::SequencerItemType::TOP_NOTE:
-                    numTopNotes++;
-                    insertedKey = true;
-                    break;
-                case NoteSequenceItem::SequencerItemType::MID_NOTE:
-                    numMidNotes++;
-                    insertedKey = true;
-                    break;
-                case NoteSequenceItem::SequencerItemType::BOT_NOTE:
-                    numBotNotes++;
-                    insertedKey = true;
-                    break;
-                default:
-                    break;
-            }
-
-            if(insertedKey) {
-                keyFrequencies[item->displayText] += 1;
-            }
-        }
-
-        updateKeyFrequencies();
-    }
+    const std::pair<std::string, int> & getKeyItemData(int frequencyRank) const;
 
     void updateKeyFrequencies() {
         keyFreqsSorted = std::vector<std::pair<std::string, int>>(keyFrequencies.begin(), keyFrequencies.end());
-        std::sort(keyFreqsSorted.begin(), keyFreqsSorted.end(), cmpSecond);
+        std::sort(keyFreqsSorted.begin(), keyFreqsSorted.end(), utils::cmpSecond);
     }
 
-    virtual int GetFrameMin() const {
-        return mFrameMin;
-    }
-    virtual int GetFrameMax() const {
-        return mFrameMax;
-    }
+    int GetFrameMin() const override { return static_cast<int>(mFrameMin); }
+    int GetFrameMax() const override { return static_cast<int>(mFrameMax); }
+    int GetItemCount() const override { return static_cast<int>(myItems.size()); }
+    int GetItemTypeCount() const override { return static_cast<int>(constants::SEQUENCER_ITEM_TYPES.size()); }
+    const char* GetItemTypeName(int typeIndex) const override { return constants::SEQUENCER_ITEM_TYPES[typeIndex].c_str(); }
+    const char* GetItemLabel(int index) const override { return ""; }
 
-    virtual int GetItemCount() const { return myItems.size(); }
-
-    virtual int GetItemTypeCount() const { return sizeof(SequencerItemTypeNames) / sizeof(char*); }
-    virtual const char* GetItemTypeName(int typeIndex) const { return SequencerItemTypeNames[typeIndex]; }
-    virtual const char* GetItemLabel(int index) const { return ""; }
-
-    virtual void Get(int index, float** start, float** end, int* type, unsigned int* color, const char** displayText)
-    {
-        auto & item = myItems[index];
-        if (color)
-            *color = 0xFFAA8080; // same color for everyone, return color based on type
-        if (start)
-            *start = &(item->absBeat);
-        if (end)
-            *end = &(item->beatEnd);
-        if(type)
-            *type = (int)(item->getItemType());
-
-        if(displayText)
-            *displayText = item->displayText.c_str();
-    }
-
-    virtual size_t GetCustomHeight(int index) { return 30; }
+    void Get(int index, float** start, float** end, int* type, unsigned int* color, const char** displayText) override;
+    size_t GetCustomHeight(int index) override { return 30; }
 };
 
 #endif // NOTESEQUENCE_HPP
